+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## Changelog
 
-### Version 1.29
+### Version 1.29.1
 
 - Logging filters added to settings
+- Fix for NX projects that expect @nrwl/cli globally
+- Rule for duplicate @ionic-native and @awesome-cordova-plugins packages
+- Project export for NX fixed
 
 ### Version 1.28